// src/app/core/interface/producto.interface.ts
export interface ProductoInterface {
  nombre: string;
  sku: string;
  cantidad: number;
  precio: number;
  createdAt: Date;
}

export interface crearProductoInterface {
  nombre: string;
  sku: string;
  cantidad: number;
  precio: number;
  distribuidor: distribuidorInterface;
}

export interface distribuidorInterface {
  nit: string;
  razonSocial: string;
  telefono: number;
  direccion: string;
}